add file with firebase objects
